+++ src/anomalib/deploy/export.py
@@ -160,6 +160,7 @@ def export_to_torch(
 def export_to_onnx(
     model: AnomalyModule,
     export_root: Path | str,
+    input_size: tuple[int, int] | None = None,
     transform: Transform | None = None,
     task: TaskType | None = None,
     export_type: ExportType = ExportType.ONNX,
@@ -169,6 +170,8 @@ def export_to_onnx(
     Args:
         model (AnomalyModule): Model to export.
         export_root (Path): Path to the root folder of the exported model.
+        input_size (tuple[int, int] | None, optional): Image size used as the input for onnx converter.
+            Defaults to None.
         transform (Transform, optional): Input transforms used for the model. If not provided, the transform is taken
             from the model.
             Defaults to ``None``.
@@ -212,14 +215,18 @@ def export_to_onnx(
     transform = transform or model.transform or model.configure_transforms()
     inference_model = InferenceModel(model=model.model, transform=transform, disable_antialias=True)
     export_root = _create_export_root(export_root, export_type)
+    input_shape = torch.zeros((1, 3, *input_size)) if input_size else torch.zeros((1, 3, 1, 1))
+    dynamic_axes = (
+        None if input_size else {"input": {0: "batch_size", 2: "height", 3: "weight"}, "output": {0: "batch_size"}}
+    )
     _write_metadata_to_json(export_root, model, task)
     onnx_path = export_root / "model.onnx"
     torch.onnx.export(
         inference_model,
-        torch.zeros((1, 3, 1, 1)).to(model.device),
+        input_shape.to(model.device),
         str(onnx_path),
         opset_version=14,
-        dynamic_axes={"input": {0: "batch_size", 2: "height", 3: "weight"}, "output": {0: "batch_size"}},
+        dynamic_axes=dynamic_axes,
         input_names=["input"],
         output_names=["output"],
     )
@@ -228,8 +235,9 @@ def export_to_onnx(
 
 
 def export_to_openvino(
-    export_root: Path | str,
     model: AnomalyModule,
+    export_root: Path | str,
+    input_size: tuple[int, int] | None = None,
     transform: Transform | None = None,
     ov_args: dict[str, Any] | None = None,
     task: TaskType | None = None,
@@ -237,8 +245,10 @@ def export_to_openvino(
     """Convert onnx model to OpenVINO IR.
 
     Args:
-        export_root (Path): Path to the export folder.
         model (AnomalyModule): AnomalyModule to export.
+        export_root (Path): Path to the export folder.
+        input_size (tuple[int, int] | None, optional): Input size of the model. Used for adding metadata to the IR.
+            Defaults to None.
         transform (Transform, optional): Input transforms used for the model. If not provided, the transform is taken
             from the model.
             Defaults to ``None``.
@@ -289,7 +299,7 @@ def export_to_openvino(
         ... )
 
     """
-    model_path = export_to_onnx(model, export_root, transform, task, ExportType.OPENVINO)
+    model_path = export_to_onnx(model, export_root, input_size, transform, task, ExportType.OPENVINO)
     ov_model_path = model_path.with_suffix(".xml")
     ov_args = {} if ov_args is None else ov_args
     if convert_model is not None and serialize is not None: